backend/requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.2.4
Django==5.2.13
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
easydict==1.13
flatbuffers==25.12.19
fonttools==4.62.1
humanfriendly==10.0
idna==3.11
ImageIO==2.37.3
insightface==0.7.3
joblib==1.5.3
kiwisolver==1.5.0
lazy-loader==0.5
matplotlib==3.10.8
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
onnx==1.21.0
onnxruntime==1.23.2
opencv-contrib-python==4.13.0.92
packaging==26.0
pillow==12.2.0
prettytable==3.17.0
protobuf==7.34.1
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
PyMySQL==1.1.2
pyparsing==3.3.2
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.3
requests==2.33.1
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
simsimd==6.5.16
six==1.17.0
sqlparse==0.5.5
stringzilla==4.6.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
wcwidth==0.6.0
loguru==0.7.3
cryptography>=42.0.0
